updates: make installable, start wrapping C-code

+++ corrcal/src/c_funcs.h
@@ -6,7 +6,7 @@ struct sparse_cov{
     int n_eig;
     int n_src;
     int n_grp;
-    int *edges;
+    long *edges;
     int isinv;
 };
 
@@ -18,6 +18,6 @@ struct sparse_cov *init_cov(
     int n_eig,
     int n_src,
     int n_grp,
-    int *edges,
+    long *edges,
     int isinv
 );
